On corrupted privateKey, generate new one
Modified the dyndnsClient.generateKeys() function to check if the privateKey is valid calling `EthCrypto.publicKeyByPrivateKey(privateKey)` inside a try/catch block. If the privateKey is not valid, a new one is generate, along the publicKey and domain.

dyndnsClient.generateKeys() is called at every VPN startup in client.js line 72
If the privateKey is corrupted a new one will be generated in that line.
On client.js line 83 `await dyndnsClient.updateIp();` is called on every VPN startup if the static IP is not set. In that case it will register the new dyndns domain.